refactor how error is triggered
