+++ frontend/src/components/main-view/grid-view/thumbnail/Thumbnail.tsx
@@ -32,7 +32,7 @@ function Thumbnail({data}: Props) {
     const onClick = () => {
         dispatch(setDetailData(data));
         dispatch(showDetails());
-        if (model !== Views.NEW && id) dispatch(selectors[model](id))
+        if (model !== Views.NEW && model !== Views.ERROR && id) dispatch(selectors[model](id))
     }
 
     return (
