+++ frontend/src/components/main-view/helpers.ts
@@ -1,6 +1,6 @@
-import {Views} from "../../interfaces/IThumbnailData";
+import {Model, Views} from "../../interfaces/IThumbnailData";
 
-export const views = (Object.keys(Views) as Array<keyof typeof Views>).slice(2)
+export const views = (Object.keys(Model) as Array<keyof typeof Model>).slice(2)
 export const images = {
     new: '',
     login: '/images/profile.svg',
@@ -8,7 +8,8 @@ export const images = {
     product: '',
     customer: '',
     supplier: '',
-    order: ''
+    order: '',
+    error:'',
 }
 export const newItemData = {
     picture:"images/add.svg",
