--- frontend/src/interfaces/IError.ts
@@ -1,5 +1,0 @@
-export interface IError {
-    status?: number;
-    statusText?: string;
-    data?:string
-}
